Further code management restructuring and introduction of ServiceLocator interface as last resort for UserIdGenerator.
